// frontend/src/shared/api/generated/models/getApiNewsParams.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * CodigoActivo.API
 * OpenAPI spec version: 1.0
 */

export type GetApiNewsParams = {
search?: string;
title?: string;
subtitle?: string;
featured?: boolean;
year?: number;
createdFrom?: string;
createdTo?: string;
page?: number;
pageSize?: number;
sort?: string;
};